refactored FinishSale to use TS
